workshops/ai-assissted-data-ingestion-with-DLT/my-dlt-pipeline/pyproject.toml
[project]
name = "my-dlt-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dlt[duckdb]>=1.22.1",
    "ibis-framework[duckdb]>=12.0.0",
    "marimo>=0.20.2",
]

[dependency-groups]
dev = [
    "dlt[workspace]>=1.22.1",
]
